TimeUntilNextThink: Suggestions from Patryk26g

+++ src/microbe_stage/systems/MicrobeAISystem.cs
@@ -187,8 +187,7 @@ public partial class MicrobeAISystem : BaseSystem<World, float>, ISpeciesMemberL
         if (ai.TimeUntilNextThink > 0)
             return;
 
-        // Random value is between -0.05 (inclusive) and 0.05 (exclusive). Thanks Desmos for helping me figure this out
-        ai.TimeUntilNextThink = Constants.MICROBE_AI_THINK_INTERVAL + (random.NextSingle() * 0.1f - 0.05f);
+        ai.TimeUntilNextThink = Constants.MICROBE_AI_THINK_INTERVAL + (float)(random.NextDouble() * 0.1 - 0.05);
 
         // This is probably pretty useless for most situations, but hopefully this doesn't eat too much
         // performance
